chore: move frames and components
