+++ packages/web/src/app/manage-club/funding/[id]/edit/page.tsx
@@ -11,9 +11,9 @@ import FlexWrapper from "@sparcs-clubs/web/common/components/FlexWrapper";
 import Modal from "@sparcs-clubs/web/common/components/Modal";
 import ConfirmModalContent from "@sparcs-clubs/web/common/components/Modal/ConfirmModalContent";
 import PageHead from "@sparcs-clubs/web/common/components/PageHead";
-import AddEvidenceFrame from "@sparcs-clubs/web/features/manage-club/funding/create/frame/AddEvidenceFrame";
-import BasicEvidenceFrame from "@sparcs-clubs/web/features/manage-club/funding/create/frame/BasicEvidenceFrame";
-import FundingInfoFrame from "@sparcs-clubs/web/features/manage-club/funding/create/frame/FundingInfoFrame";
+import AddEvidenceFrame from "@sparcs-clubs/web/features/manage-club/funding/frame/AddEvidenceFrame";
+import BasicEvidenceFrame from "@sparcs-clubs/web/features/manage-club/funding/frame/BasicEvidenceFrame";
+import FundingInfoFrame from "@sparcs-clubs/web/features/manage-club/funding/frame/FundingInfoFrame";
 import { FundingInterface } from "@sparcs-clubs/web/features/manage-club/funding/types/funding";
 import mockFundingDetail from "@sparcs-clubs/web/features/manage-club/service/_mock/mockFundingDetail";
 import { formatDotDate } from "@sparcs-clubs/web/utils/Date/formateDate";
